made a layout of the match (item_match.xml) made adapter and recyclerview

+++ app/src/main/java/course/apps/footballmatches/pojo/FixtureInfo.kt
@@ -24,7 +24,7 @@ data class FixtureInfo(
 
     @SerializedName("timestamp")
     @Expose
-    val timestamp: Int? = null,
+    val timestamp: Long? = null,
 
     @SerializedName("status")
     @Expose
